refactor(hubspot): move connectors common folder
