+++ apps/hubspot/src/connectors/hubspot/common/error.ts
