+++ apps/hubspot/src/inngest/functions/token/refresh-token.test.ts
@@ -6,7 +6,7 @@ import { db } from '@/database/client';
 import { organisationsTable } from '@/database/schema';
 import * as authConnector from '@/connectors/hubspot/auth';
 import { encrypt, decrypt } from '@/common/crypto';
-import { HubspotError } from '@/connectors/common/error';
+import { HubspotError } from '@/connectors/hubspot/common/error';
 import { refreshToken } from './refresh-token';
 
 const timeZone = 'us/eastern';
